Extract form types from types.rs to fix source file length check (Ref #92)
